CLI: move create-package steps
