--- packages/frontity/src/steps/create-package/__tests__/steps.tests.ts
@@ -1,3 +1,0 @@
-describe("steps", () => {
-  test.todo("add tests");
-});
